feat: refactor storage and widget hooks; enhance type handling, improve widget retrieval logic, and consolidate bookmark types
