+++ src/chrome/storage.ts
@@ -5,9 +5,12 @@ export const syncSet = async (key: StorageKey, value: object | string): Promise<
   chrome.storage.sync.set({ [key]: value });
 };
 
-export const syncGet = async <T extends object | string>(key: StorageKey): Promise<T> => {
+export const syncGet = async <T extends object | string>(key: StorageKey): Promise<T | undefined> => {
   return new Promise((resolve) => {
     chrome.storage.sync.get(key, (result) => {
+      if (!result[key]) {
+        resolve(undefined);
+      }
       resolve(result[key]);
     });
   });
