--- src/types/Bookmarks.ts
@@ -1,4 +1,0 @@
-export type WidgetBookmarkType = Omit<chrome.bookmarks.BookmarkTreeNode, 'children'> & {
-  imageUrl?: string;
-  children?: WidgetBookmarkType[] | undefined;
-};
