+++ src/types/Widget.ts
@@ -6,6 +6,13 @@ export type WidgetType<T> = {
   data: T;
 };
 
+export type WidgetBookmarkType = Omit<chrome.bookmarks.BookmarkTreeNode, 'children'> & {
+  imageUrl?: string;
+  children?: WidgetBookmarkType[] | undefined;
+};
+
 export type ClockWidgetType = {
   format: string;
 };
+
+export type CustomWidgetType = WidgetBookmarkType | ClockWidgetType;
